requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
defusedxml==0.7.1
Django==5.2.6
django-cors-headers==4.8.0
django-dotenv==1.4.2
django-templated-mail==1.1.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
google-auth==2.40.3
google-genai==1.36.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
oauthlib==3.3.1
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.4
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.11.0
